requirements.txt
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dpcpp-cpp-rt==2023.0.0
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.38.0
impi-rt==2021.8.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
intel-cmplr-lib-rt==2023.0.0
intel-cmplr-lic-rt==2023.0.0
intel-fortran-rt==2023.0.0
intel-numpy==1.21.4
intel-opencl-rt==2023.0.0
intel-openmp==2023.0.0
intel-scipy==1.7.3
itsdangerous==2.1.2
jax==0.4.4
jaxlib==0.4.4
Jinja2==3.1.2
julia==0.6.0
kiwisolver==1.4.4
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.7.0
mkl==2023.0.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mkl-umath==0.1.1
numba==0.56.4
numpy==1.21.4
numpy-stl==3.0.0
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-utils==3.4.5
scipy==1.7.3
six==1.16.0
tbb==2021.8.0
tbb4py==2021.8.0
Werkzeug==2.2.3
zipp==3.15.0
